change int to long

+++ Main/DTOs/Room/PlaceBetsDTO.cs
@@ -6,6 +6,6 @@ namespace _3D_Tim_backend.DTOs
     public record class PlaceBetsDTO
     (
         [Required][Range(0, 9999)] int RoomId,
-        [Required] Dictionary<string, int> BetSidesWithAmount
+        [Required] Dictionary<string, long> BetSidesWithAmount
         );
 }
